+++ index.html
@@ -71,7 +71,7 @@
             <form class="input-bar-wrap">
               <fieldset class="date">
                 <label for="date-input">일자</label>
-                <input type="number" id="date-input" pattern="[0-9]{8}" title="'yyyymmdd'와 같은 8자리 숫자" required>
+                <input type="text" id="date-input" pattern="[0-9]{8}" title="'yyyymmdd'와 같은 8자리 숫자" required>
               </fieldset>
               <div class="dividing-line"></div>
               <fieldset class="price">
