+++ src/js/events/tabs/index.js
@@ -3,7 +3,6 @@ import { $, $All } from "../../utils/dom.js";
 export const tabsEventHandler = () => {
   const tabButtons = $All(".tab-button");
   const tabContents = $All(".tab-content");
-  // const buttonPaths = $All(".tab-button path");
 
   tabButtons.forEach((button, index) => {
     button.addEventListener("click", function (event) {
@@ -13,18 +12,7 @@ export const tabsEventHandler = () => {
 
       tabContents[index].classList.add("active");
 
-      /* 버튼 색상 */
       changeBtnColor({ index, event });
-      // buttonPaths.forEach((path) => {
-      //   path.setAttribute("stroke", "#A79FCB");
-      // });
-
-      // if (event.currentTarget === tabButtons[index]) {
-      //   const paths = event.currentTarget.querySelectorAll("path");
-      //   paths.forEach((path) => {
-      //     path.setAttribute("stroke", "#FCFCFC");
-      //   });
-      // }
     });
   });
 };
