+++ src/js/init/initListTotalInfo.js
@@ -1,0 +1,7 @@
+import { renderListTotalLength } from "../view/renderListTotalLength.js";
+import { renderTotalIncomeExpenditure } from "../utils/renderTotalIncomeExpenditure.js";
+
+export const initListTotalInfo = () => {
+  renderTotalIncomeExpenditure()
+  renderListTotalLength();
+};
