--- src/js/init/initListTotalLength.js
@@ -1,5 +1,0 @@
-import { renderListTotalLength } from "../utils/renderListTotalLength.js";
-
-export const initListTotalLength = () => {
-  renderListTotalLength();
-};
