--- src/js/init/initTotalIncomeExpenditure.js
@@ -1,5 +1,0 @@
-import { renderTotalIncomeExpenditure } from "../utils/renderTotalIncomeExpenditure.js";
-
-export const initTotalIncomeExpenditure = () => {
-  renderTotalIncomeExpenditure();
-};
