+++ src/js/utils/addUpAllIncomeExpenditure.js
@@ -1,5 +1,5 @@
 //총 건수, 총 합계 => 해당 월의 총 건수와 총 합계로 바뀌어야 함
-const addAllIncome = () => {
+const addUpAllIncome = () => {
   //키를 가져옴, 키로 각 항목을 배열로 만들어줌
   return Object.keys(localStorage)
     .map((el) => JSON.parse(localStorage.getItem(el)))
@@ -9,7 +9,7 @@ const addAllIncome = () => {
     }, 0);
 };
 //함수를 따로 놓지 않을 수 있지 않을가...
-const addAllExpenditure = () => {
+const addUpAllExpenditure = () => {
   return Object.keys(localStorage)
     .map((el) => JSON.parse(localStorage.getItem(el)))
     .filter((el) => el.type === false)
@@ -18,4 +18,4 @@ const addAllExpenditure = () => {
     }, 0);
 };
 
-export { addAllIncome, addAllExpenditure };
+export { addUpAllIncome, addUpAllExpenditure };
