+++ src/js/view/renderListTotalLength.js
@@ -1,5 +1,5 @@
-import { $ } from "./dom.js";
-import { countListTotalLength } from "./countListTotalLength.js";
+import { $ } from "../utils/dom.js";
+import { countListTotalLength } from "../utils/countListTotalLength.js";
 
 export const renderListTotalLength = () => {
   $(".info-total-count").innerText = countListTotalLength() + "건";
